refactor: move Sysinfo to Admin namespace
